# src/static/swagger.json
{
    "openapi" : "3.0.0",
    "info" : {
            "decription" : "API desc",
            "version" : "1.0.0",
            "title" : "Open API for Mexico's Mortality Open Data (2012-2020)",
            "contact" : {
                "name" : "Roberto Mares Ceballos",
                "email" : "roberto.mares.dev@gmail.com"
            },
            "license":{
                "name" : "OPEN SOURCE",
                "url" : ""
            }
    },
    "servers": [
        {
          "url": "/"
        }
    ],
    "paths": {
        "/mortality/v.1/mortality-state" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers of a State from 2012 - 2020",
                "parameters": [
                    {
                        "in": "query",
                        "name": "state",
                        "required": true,
                        "description" : "The number of the Entity of ocurrence",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/state"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/mortality-state-year" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers of a State and Year",
                "parameters": [
                    {
                        "in": "query",
                        "name": "state",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Entity",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/state"
                        }
                    },
                    {
                        "in": "query",
                        "name": "year",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of State",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/year"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/mortality-scholarship-year" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Scholarship and Year",
                "parameters": [
                    {
                        "in": "query",
                        "name": "scholarship",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Scholarship",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/scholarship"
                        }
                    },
                    {
                        "in": "query",
                        "name": "year",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Year",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/year"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/mortality-scholarship-state" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Scholarship and State",
                "parameters": [
                    {
                        "in": "query",
                        "name": "scholarship",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Scholarship",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/scholarship"
                        }
                    },
                    {
                        "in": "query",
                        "name": "state",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of State",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/state"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/mortality-sex-year" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Sex and Year",
                "parameters": [
                    {
                        "in": "query",
                        "name": "sex",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Sex",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/sex"
                        }
                    },
                    {
                        "in": "query",
                        "name": "year",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Year",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/year"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/mortality-sex-state" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Sex and State",
                "parameters": [
                    {
                        "in": "query",
                        "name": "sex",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Sex",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/sex"
                        }
                    },
                    {
                        "in": "query",
                        "name": "state",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of State",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/state"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/mortality-medical-year" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Medical Attention and Year",
                "parameters": [
                    {
                        "in": "query",
                        "name": "medical",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Sex",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/medical"
                        }
                    },
                    {
                        "in": "query",
                        "name": "year",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Year",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/year"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/mortality-age-range" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Age range",
                "parameters": [
                    {
                        "in": "query",
                        "name": "age",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Age range",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/age"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/mortality-age-range-year" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Age range and specific Year",
                "parameters": [
                    {
                        "in": "query",
                        "name": "age",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Age range",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/age"
                        }
                    },
                    {
                        "in": "query",
                        "name": "year",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Year",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/year"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/top-morbidity" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Top 10 Morbiditys, from Lista Mex. From 2012-2020.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/top-morbidity-year" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Top 10 Morbiditys from Lista Mex, and specific Year.",
                "parameters": [
                    {
                        "in": "query",
                        "name": "year",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Year",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/year"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/top-mortality-states" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Top 10 States, from 2012-2020.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        },
        "/mortality/v.1/top-mortality-states-year" : {
            "get": {
                "tags": [
                  "API Requests"
                ],
                "summary": "Gets the mortality registers based on Top 10 States specific Year (2012-2020).",
                "parameters": [
                    {
                        "in": "query",
                        "name": "year",
                        "type" : "integer",
                        "required": true,
                        "description" : "The id of Year",
                        "schema" :{
                            "$ref" : "#/components/schemas/parameters/properties/year"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/mortality_register"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    }
                }
            }
        }
    },
...
}